refactor: :truck: Split SearchIcon into FontAwesomePng
Issue: #181
